feat: Crew score Integer에서 Float으로 변경
- 기본 스코어가 Float이기 때문에 맞춰주기 위해
- CrewInfo에 스코어도 함께 반환되도록 추가

+++ project/back/health/src/main/java/com/ssafy/health/domain/crew/dto/response/CrewDetailResponseDto.java
@@ -17,7 +17,7 @@ public class CrewDetailResponseDto {
     private Long totalBattlesCount;
     private Long winCount;
     private Float averageAge;
-    private Integer basicScore;
-    private Integer activityScore;
+    private Float basicScore;
+    private Float activityScore;
     private CrewRole role;
 }
